Improves book instance functionality for library.
Adds functionality to add Book instances to cards in HTML.
Adds event listeners to be able to click on book to open/close for
editing.
Improves cover background images on Book instances.
Adds transtion of book opening and closing instead of animation.
Adds form as HTML to each Book instance to be able to edit properties.